backend/requirements.txt
Flask==2.0.2
psycopg2-binary==2.9.1
Flask-Cors==3.0.10
Werkzeug==2.0
python-dotenv==0.19.0
